requirements.txt
Quart == 0.10.0
RPi.GPIO == 0.7.0